src/domain/dtos/registerRequest.dto.ts
import {
  IsEmail,
  IsNotEmpty,
  IsString,
  IsStrongPassword,
} from 'class-validator';

export class RegisterDTO {
  @IsString()
  @IsNotEmpty()
  name: string;

  @IsString()
  @IsEmail()
  @IsNotEmpty()
  email: string;

  @IsString()
  @IsNotEmpty()
  surname: string;

  @IsString()
  @IsNotEmpty()
  @IsStrongPassword(
    {
      minLength: 6,
      minUppercase: 1,
      minLowercase: 1,
      minSymbols: 0,
      minNumbers: 1,
    },
    {
      message:
        'Password not strong enough. It must have more than 6 characters, 1 uppercase, 1 lowercase and 1 number',
    },
  )
  password: string;
}
